potato_counter/requirements.txt
PyQt5>=5.15.0
opencv-python>=4.8.0
numpy>=1.24.0
PyYAML>=6.0
pydantic>=2.0.0
